make python the process controlled by docker
pygame has to cleanup the framebuffer otherwise the next time it runs it
gets stuck at display.set_mode(). This was the case before this commit
because systemd was on and the SIGTERM signal from Docker wasn't
forwarded to the python process. Now when a new update is pushed, resin
will send SIGTERM to the python process which do the right cleanup.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>